Move include/algorithm/algorithm.h to include/
